% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCorrelationImageProblem.R
\name{makeCorrelationImageProblem}
\alias{makeCorrelationImageProblem}
\title{Correlation Image Problem}
\usage{
makeCorrelationImageProblem(n = 5000L, resolution = c(32, 32),
  points.per.image = 100, pointsize = 2, create.task = TRUE)
}
\arguments{
\item{n}{[\code{integer(1)}]\cr
Number of observations to create.
Default is \code{5000L}.}

\item{resolution}{[\code{integer(2)}]\cr
Resolution of the image. This defines the number of features.
Default is \code{c(32, 32)}, which creates \code{1024} features.}

\item{points.per.image}{[\code{integer(1)}]\cr
How man points should be generated per image.
Default is \code{100}.}

\item{pointsize}{[\code{integer(1)}]\cr
Pointsize of intepreted big points (1/72 inch). See \code{\link{bmp}}.
Default is \code{2}.}

\item{create.task}{[\code{logical(1)}]\cr
Should a task be created? If \code{TRUE} a \pkg{mlr} regression task (see \code{\link[mlr]{makeRegrTask}}
is returned. Otherwise a \code{data.frame} is returned.}
}
\value{
\code{RegrTask} | \code{data.frame}.
}
\description{
This benchmark is motivated by \url{http://guessthecorrelation.com/}. The task is to
assess the correlation between two variables visually by a scatterplot of sampled points.
Each observation is generated as follows:
\itemize{
 \item{A random correlation value is drawn uniformly between -1 and 1.}
 \item{\code{points.per.image} points are sampled from a bivariate normal distribution with zero mean, sampled correlation and a marginal variance of 1.}
 \item{The observations are scaled between 0 and 1.}
 \item{An image of the data is generated and saved in a given resolution and pointsize.}
 \item{The values of the pixels are returned as a vector with the correlation as target.}
}
}
\examples{
 data = makeCorrelationImageProblem(n = 20, create.task = FALSE)
 image(matrix(as.numeric(data[1, -1]), ncol = 32))
}
