% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFizzbuzzProblem.R
\name{makeFizzbuzzProblem}
\alias{makeFizzbuzzProblem}
\title{Fizzbuzz Classification Problem}
\usage{
makeFizzbuzzProblem(n = 10^4, create.task = TRUE)
}
\arguments{
\item{n}{[\code{integer(1)}]\cr
Number of observations to create.
Default is \code{5000L}.}

\item{create.task}{[\code{logical(1)}]\cr
Should a task be created? If \code{TRUE} a \pkg{mlr} classification task (see \code{\link[mlr]{makeClassifTask}}
is returned. Otherwise a \code{data.frame} is returned.}
}
\value{
\code{ClassifTask} | \code{data.frame}.
}
\description{
This problem is inspired by this blog post \url{http://joelgrus.com/2016/05/23/fizz-buzz-in-tensorflow/}.
Obviously this is not a serious benchmarking problem, but it isn't totally trivial to solve.
Each observation is one number encoded in 32-bit binary. The target if
\itemize{
 \item{\code{"fizz"} if the number is dividable by 3.}
 \item{\code{"buzz"} if the number is dividable by 5.}
 \item{\code{"fizzbuzz"} if the number is dividable by 15.}
 \item{\code{"number"} otherwise.}
}
}
\examples{
 makeFizzbuzzProblem(1, create.task = FALSE)
}
