% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNotNotMnist.R
\name{makeNotNotMnistProblem}
\alias{makeNotNotMnistProblem}
\title{MNIST like problem.}
\usage{
makeNotNotMnistProblem(n = 5^4, sample.font = TRUE, random.rotation = 0,
  position.range = c(0.3, 0.7), size.range = c(15, 50), letters = LETTERS,
  resolution = c(32, 32), pointsize = 2, create.task = TRUE)
}
\arguments{
\item{n}{[\code{integer(1)}]\cr
Number of observations to create.
Default is \code{5000L}.}

\item{sample.font}{[\code{logical(1)}]\cr
Should every image be sampled from a different font. Fonts are sampled from \code{\link{Hershey}}.
Default is \code{TRUE}.}

\item{random.rotation}{[\code{numeric(1)}]\cr
Should the rotation of the letters/numbers be randomized. Value is the maximum degree of rotation.
Default is 0, which is no rotation at all.}

\item{position.range}{[\code{numeric(2}]\cr
Randomize position of the letter/number. First number is the lower bound for x/y value,
second number the upper bound. Minimal/maximal values of the canvas are 0 and 1.
Note that it is possible that some element of the letters/number might be out of the bound,
depending on the range and size.
Default is \code{c(0.3, 0.7)}}

\item{size.range}{[\code{numeric(2)}]\cr
Size of the letters/numbers, is sampled between first and second value. This is the
\code{cex} value of the plot.
Default is \code{c(15, 50)}.}

\item{letters}{[\code{numeric() | character()}]\cr
Letters/Numbers to sample from. Can be anything that can be passed to \code{text()}.
Default is \code{LETTERS}.}

\item{resolution}{[\code{integer(2)}]\cr
Resolution of the image. This defines the number of features.
Default is \code{c(32, 32)}, which creates \code{1024} features.}

\item{pointsize}{[\code{integer(1)}]\cr
Pointsize of intepreted big points (1/72 inch). See \code{\link{bmp}}.
Default is \code{2}.}

\item{create.task}{[\code{logical(1)}]\cr
Should a task be created? If \code{TRUE} a \pkg{mlr} classification task (see \code{\link[mlr]{makeClassifTask}}
is returned. Otherwise a \code{data.frame} is returned.}
}
\value{
\code{ClassifTask} | \code{data.frame}.
}
\description{
Generate images of letters or numbers in various ways. Randomized size, position, rotations
and fonts are available to set the difficulty of the problem.
}
\examples{
 data = makeNotNotMnistProblem(10, create.task = FALSE)
 image(matrix(as.numeric(data[1, -1]), ncol = 32))
}
